+++ pkg/deployment/resources/inspector/sa_anonymous.go
@@ -35,7 +35,7 @@ func (p *serviceAccountsInspector) Anonymous(gvk schema.GroupVersionKind) (anony
 			if p.v1 == nil || p.v1.err != nil {
 				return nil, false
 			}
-			return &serviceAccountsInspectorAnonymousV1{i: p.v1}, true
+			return &serviceAccountsInspectorAnonymousV1{i: p.state}, true
 		}
 	}